requirements.txt
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
SQLAlchemy==2.0.44
pyodbc==5.3.0